Health bar added
- Added health bar for Mage and Knight in PvP mode

+++ Assets/Scripts/Control.cs
@@ -31,7 +31,7 @@ public class Control : MonoBehaviour
     LayerMask enemyLayer;
 
     [SerializeField]
-    int damage = 20;
+    float damage = 20f;
 
     bool defend = false;
