+++ Assets/Scripts/EnemyAttack.cs
@@ -24,7 +24,7 @@ public class EnemyAttack : MonoBehaviour
     LayerMask enemyLayer;
 
     [SerializeField]
-    int damage = 20;
+    float damage = 20f;
 
     Timer timer;
